Fix label dtype to int64 for CrossEntropyLoss compatibility

+++ bonsai/functional/create_data.py
@@ -60,7 +60,7 @@ def process_split(
         # Cast to correct dtypes
         tokenized = tokenized.select(
             pl.col("subject_id").cast(pl.Int64),
-            pl.col("code").cast(pl.Int32),
+            pl.col("code").cast(pl.Int64),
             pl.col("age").cast(pl.Float16),
             pl.col("abspos").cast(pl.Float32),
             pl.col("segment").cast(pl.Int32),
